Upload optimization

+++ app/src/main/java/com/infomaniak/drive/data/models/upload/UploadSegment.kt
@@ -25,7 +25,7 @@ data class UploadSegment(
     val status: ChunkStatus,
     @SerializedName("created_at")
     val createdAt: Date,
-    val size: Long,
+    val size: Int,
     /** Chunk uploaded md5 hash, computed */
     val hash: String
 ) {
